+++ src/cyclops/Thread.h
@@ -4,7 +4,7 @@
 #include <algorithm>
 #include <thread>
 #include <mutex>
-#include "tinythread/tinythread.h" 
+#include "tinythread/tinythread.h"
 
 #if defined(_WIN32) || defined(__WIN32__) || defined(__WINDOWS__) || defined(WIN_BUILD)
     #define USE_TTHREAD
@@ -32,17 +32,17 @@ namespace threading {
     typedef threading::std_thread DefaultThreadType;
 #endif
 
-namespace threading {   
+namespace threading {
 namespace tthread {
-    
+
     template <typename ItType, typename Function>
     struct for_each_arguments {
         ItType begin;
         ItType end;
         Function function;
-    
-        for_each_arguments(ItType begin, ItType end, Function function) 
-            : begin(begin), end(end), function(function) { }           
+
+        for_each_arguments(ItType begin, ItType end, Function function)
+            : begin(begin), end(end), function(function) { }
     };
 
     template <typename ItType, typename Function>
@@ -52,7 +52,7 @@ namespace tthread {
            auto args = bsccs::unique_ptr<ArgType>( // To avoid leaks
                 static_cast<ArgType*>(a)
             );
-            
+
             std::for_each(args->begin, args->end, args->function);
        } catch (...) {
        }
@@ -63,76 +63,76 @@ namespace tthread {
 template <typename InputIt>
 struct TaskScheduler {
 
-	TaskScheduler(InputIt begin, InputIt end, const size_t nThreads) 
-	   : begin(begin), end(end), 
+	TaskScheduler(InputIt begin, InputIt end, const size_t nThreads)
+	   : begin(begin), end(end),
 	     taskCount(std::distance(begin, end)),
-	     nThreads(std::min(nThreads, taskCount)), 
+	     nThreads(std::min(nThreads, taskCount)),
 	     chunkSize(
 	     	taskCount / nThreads + (taskCount % nThreads != 0)
 	     ) { }
-         
+
     template <typename UnaryFunction>
     UnaryFunction execute(UnaryFunction function) {
-        return execute(function, DefaultThreadType());   
+        return execute(function, DefaultThreadType());
     }
-	     	
+
 	size_t getThreadIndex(size_t i) {
 		return nThreads == 1 ? 0 :
 			i / chunkSize;
-	}	
-	
+	}
+
 	size_t getChunkSize() const { return chunkSize; }
-	
+
 	int getThreadCount() const { return nThreads; }
-	
+
 private:
 
 #ifdef USE_TTHREAD
 	template <typename UnaryFunction>
-	UnaryFunction execute(UnaryFunction function, threading::tthread_thread) {	
-			
+	UnaryFunction execute(UnaryFunction function, threading::tthread_thread) {
+
         std::vector<tthread::thread*> workers;
 		size_t start = 0;
-		for (int i = 0; i < nThreads - 1 && begin + start + chunkSize < end; ++i, start += chunkSize) {
-		
+		for (size_t i = 0; i < nThreads - 1 && begin + start + chunkSize < end; ++i, start += chunkSize) {
+
             workers.emplace_back(new tthread::thread(
-                threading::tthread::for_each<InputIt,UnaryFunction>, 
+                threading::tthread::for_each<InputIt,UnaryFunction>,
                 new threading::tthread::for_each_arguments<InputIt, UnaryFunction>(
-                begin + start, 
-                begin + start + chunkSize, 
-                function)));			            
+                begin + start,
+                begin + start + chunkSize,
+                function)));
 		}
-				
+
 		auto rtn = std::for_each(begin + start, end, function);
-		for (int i = 0; i < workers.size(); ++i) {
+		for (size_t i = 0; i < workers.size(); ++i) {
 			workers[i]->join();
 			delete workers[i];
 		}
-		
-		return rtn;	
-	}	
+
+		return rtn;
+	}
 #else
     template <typename UnaryFunction>
-	UnaryFunction execute(UnaryFunction function, threading::std_thread) {	
+	UnaryFunction execute(UnaryFunction function, threading::std_thread) {
 
-		std::vector<std::thread> workers;		
+		std::vector<std::thread> workers;
 		size_t start = 0;
 		for (int i = 0; i < nThreads - 1 && begin + start + chunkSize < end; ++i, start += chunkSize) {
 
             workers.emplace_back(std::thread(
 				std::for_each<InputIt, UnaryFunction>,
-				begin + start, 
-				begin + start + chunkSize, 
-				function));				
+				begin + start,
+				begin + start + chunkSize,
+				function));
 		}
-		
+
 		auto rtn = std::for_each(begin + start, end, function);
 		for (int i = 0; i < workers.size(); ++i) {
 			workers[i].join();
-		}				
-		return rtn;	
+		}
+		return rtn;
 	}  // TODO Remove code-duplication between std::thread and tthread::thread versions
-#endif    
+#endif
 
 	const InputIt begin;
 	const InputIt end;
